Georeferencing: Display georeferencing information on a map
This patch adds a small map to the georeferencing window which shows
the precise extent of the map on OpenStreetMap. This is a sanity check
to make sure the georeferencing information for the map is correct.
There are quite a lot of maps uploaded which have the wrong
georeferencing information or the world file has been given the wrong
CRS.